Deals successfully rendering to page
