+++ package.json
@@ -8,12 +8,14 @@
     "node": ">=6.13.0"
   },
   "scripts": {
+    "start": "nodemon server/index.js",
     "start-webpack": "webpack -d --watch",
-    "seed-mongo": "node database/index.js"
+    "seed-mongo": "node database/seed.js"
   },
   "dependencies": {
     "axios": "^0.18.0",
     "babel": "^6.23.0",
+    "express": "^4.16.3",
     "faker": "^4.1.0",
     "mongoose": "^5.2.13",
     "npx": "^10.2.0",
